+++ src/whatsapp/dto/chat.dto.ts
@@ -26,8 +26,11 @@ export class NumberBusiness {
   message?: string;
   description?: string;
   email?: string;
-  website?: string[];
+  websites?: string[];
   address?: string;
+  about?: string;
+  vertical?: string;
+  profilehandle?: string; 
 }
 
 export class ProfileNameDto {
