+++ src/whatsapp/routers/sendMessage.router.ts
@@ -13,6 +13,7 @@ import {
   statusMessageSchema,
   stickerMessageSchema,
   textMessageSchema,
+  templateMessageSchema,
 } from '../../validate/validate.schema';
 import { RouterBroker } from '../abstract/abstract.router';
 import {
@@ -27,6 +28,7 @@ import {
   SendStatusDto,
   SendStickerDto,
   SendTextDto,
+  SendTemplateDto,
 } from '../dto/sendMessage.dto';
 import { sendMessageController } from '../whatsapp.module';
 import { HttpStatus } from './index.router';
@@ -129,6 +131,22 @@ export class MessageRouter extends RouterBroker {
           schema: listMessageSchema,
           ClassRef: SendListDto,
           execute: (instance, data) => sendMessageController.sendList(instance, data),
+        });
+
+        return res.status(HttpStatus.CREATED).json(response);
+      })
+      .post(this.routerPath('sendTemplate'), ...guards, async (req, res) => {
+        logger.verbose('request received in sendTemplate');
+        logger.verbose('request body: ');
+        logger.verbose(req.body);
+
+        logger.verbose('request query: ');
+        logger.verbose(req.query);
+        const response = await this.dataValidate<SendTemplateDto>({
+          request: req,
+          schema: templateMessageSchema,
+          ClassRef: SendTemplateDto,
+          execute: (instance, data) => sendMessageController.sendTemplate(instance, data),
         });
 
         return res.status(HttpStatus.CREATED).json(response);
