+++ src/whatsapp/routers/webhook.router.ts
@@ -7,11 +7,12 @@ import { InstanceDto } from '../dto/instance.dto';
 import { WebhookDto } from '../dto/webhook.dto';
 import { webhookController } from '../whatsapp.module';
 import { HttpStatus } from './index.router';
+import { WABussiness, ConfigService } from '../../config/env.config';
 
 const logger = new Logger('WebhookRouter');
 
 export class WebhookRouter extends RouterBroker {
-  constructor(...guards: RequestHandler[]) {
+  constructor(readonly configService: ConfigService, ...guards: RequestHandler[]) {
     super();
     this.router
       .post(this.routerPath('set'), ...guards, async (req, res) => {
@@ -30,6 +31,32 @@ export class WebhookRouter extends RouterBroker {
 
         res.status(HttpStatus.CREATED).json(response);
       })
+      .post(this.routerPath('whatsapp'), async (req, res) => {
+        logger.verbose('request received in findChatwoot');
+        logger.verbose('request body: ');
+        logger.verbose(req.body);
+
+        logger.verbose('request query: ');
+        logger.verbose(req.query);
+        const response = await this.dataValidate<InstanceDto>({
+          request: req,
+          schema: instanceNameSchema,
+          ClassRef: InstanceDto,
+          execute: (instance, data) => webhookController.receiveWebhook(instance, data),
+        });
+
+        res.status(HttpStatus.OK).json(response);
+      })
+      .get(this.routerPath('whatsapp'), async (req, res) => {
+        logger.verbose('request received in webhook');
+        logger.verbose('request query: ');
+        logger.verbose(req.query);
+        if (req.query['hub.verify_token'] === this.configService.get<WABussiness>('WABUSSINESS').ACESS_TOKEN)
+          res.send(req.query['hub.challenge']);
+        else
+          res.send('Error, wrong validation token');    
+        logger.verbose('Error, wrong validation token');
+        })
       .get(this.routerPath('find'), ...guards, async (req, res) => {
         logger.verbose('request received in findWebhook');
         logger.verbose('request body: ');
